requirements.txt
python-telegram-bot==21.10
openai-whisper==20231117
